refactor: single-source the OmniJS vm sandbox fixture across AST tests
/code-review high finding on #196: the vm stub (Task.Status enum, stub task
factory, script runners) existed as four hand-rolled copies that had already
drifted — project-root-exclusion declared a non-existent Task.Status.Active
member, and both omn-130 copies omitted Completed.

New tests/unit/contracts/ast/omnijs-vm-fixture.ts is the single source
(real 7-member status enum verbatim); filter-projection-agreement,
omn-130-read-honesty, and project-root-exclusion now consume it via thin
local aliases so call sites are unchanged.
